requirements.txt
python-telegram-bot
bidict
